(PDK-1099) Merge Puppet::Util::Windows into PDK namespace
