+++ lib/pdk/util/windows/api_types.rb
@@ -1,0 +1,57 @@
+require 'ffi'
+require 'pdk/util/windows/string'
+
+module PDK::Util::Windows::APITypes
+  module ::FFI
+    WIN32_FALSE = 0
+
+    # standard Win32 error codes
+    ERROR_SUCCESS = 0
+  end
+
+  class ::FFI::Pointer
+    def self.from_string_to_wide_string(str, &_block)
+      str = PDK::Util::Windows::String.wide_string(str)
+      FFI::MemoryPointer.new(:byte, str.bytesize) do |ptr|
+        # uchar here is synonymous with byte
+        ptr.put_array_of_uchar(0, str.bytes.to_a)
+
+        yield ptr
+      end
+
+      # ptr has already had free called, so nothing to return
+      nil
+    end
+
+    def read_wide_string(char_length, dst_encoding = Encoding::UTF_8, encode_options = {})
+      # char_length is number of wide chars (typically excluding NULLs), *not* bytes
+      str = get_bytes(0, char_length * 2).force_encoding('UTF-16LE')
+      str.encode(dst_encoding, str.encoding, encode_options)
+    rescue StandardError => e
+      PDK.logger.debug _('Unable to convert value %{string} to encoding %{encoding} due to %{error}') % {
+        string:   str.dump,
+        encoding: dst_encoding,
+        error:    e.inspect,
+      }
+      raise
+    end
+  end
+
+  # FFI Types
+  # https://github.com/ffi/ffi/wiki/Types
+
+  # Windows - Common Data Types
+  # https://msdn.microsoft.com/en-us/library/cc230309.aspx
+
+  # Windows Data Types
+  # https://msdn.microsoft.com/en-us/library/windows/desktop/aa383751(v=vs.85).aspx
+
+  FFI.typedef :uint32, :dword
+  # buffer_inout is similar to pointer (platform specific), but optimized for buffers
+  FFI.typedef :buffer_inout, :lpwstr
+  # buffer_in is similar to pointer (platform specific), but optimized for CONST read only buffers
+  FFI.typedef :buffer_in, :lpcwstr
+  # 8 bits per byte
+  FFI.typedef :uchar, :byte
+  FFI.typedef :uint16, :wchar
+end
